REFACTOR: auth/user 컨트롤러 분리
